Port ValueMesh to Rust
Summary:
We have both a `ValueMesh` defined in Rust in monarch_hyperactor as well another Python wrapper over it defined in monarch/python/monarch/_src/actor/actor_mesh.py. 

The Python wrapper exists to provide the `MeshTrait` mixin. In the previous diff we ported the `MeshTrait` methods to Rust, so now we can make the Rust `ValueMesh` implement this trait and eliminate the Python wrapper

Differential Revision: D89965880